move alibi test into prefix prefill tests
